Add blockinfo feature to system contract.
The blockinfo feature adds metadata (block height and timestamp) for
each block into a table within the system contract. Only a fixed size
rolling window of entries is maintained in that table. Currently, that
size is hardcoded to 10.

This table is in a well-defined format to allow other contracts to
reliably read from it. A function, get_latest_block_batch_info, is
provided in the `eosio.system/block_info.hpp` header file as a
convenient way to access the information in the table from within other
contracts.

That function is versatile in that it enables getting the metadata for
a block specified by a particular block height (assuming it is still
tracked in the table), or getting the metadata for the last recorded
block, or even getting the metadata for the last batch of blocks in
which the sequence of blocks is partitioned into contiguous fixed size
chunks of contiguous blocks.

The tests of the feature demonstrate the different ways of querying
that block metadata as well as some of the error conditions to watch
out for.

This commit also does some code cleanup and fixes a few spelling issues.